Refactor audit system and icon components into modular structure
