+++ frontend-react/src/components/icons/actions.ts
@@ -1,0 +1,124 @@
+/**
+ * Action Icons
+ * 
+ * Icons used for user actions and interactions.
+ */
+
+import {
+  PlusOutlined,
+  EditOutlined,
+  DeleteOutlined,
+  SaveOutlined,
+  CloseOutlined,
+  CheckOutlined,
+  SendOutlined,
+  DownloadOutlined,
+  UploadOutlined,
+  EyeOutlined,
+  EyeInvisibleOutlined,
+  LockOutlined,
+  UnlockOutlined,
+  ReloadOutlined,
+  SyncOutlined,
+  RedoOutlined,
+  UndoOutlined,
+  RollbackOutlined,
+  ForwardOutlined,
+  BackwardOutlined,
+  PlayCircleOutlined,
+  PauseCircleOutlined,
+  StopOutlined,
+  StepForwardOutlined,
+  StepBackwardOutlined,
+  FastForwardOutlined,
+  FastBackwardOutlined,
+  ShrinkOutlined,
+  ArrowsAltOutlined,
+  FullscreenOutlined,
+  FullscreenExitOutlined,
+  ZoomInOutlined,
+  ZoomOutOutlined,
+  CompressOutlined,
+  ExpandOutlined,
+  SwapOutlined,
+  SwapLeftOutlined,
+  SwapRightOutlined,
+  SortAscendingOutlined,
+  SortDescendingOutlined,
+  FilterOutlined,
+  FunnelPlotOutlined,
+  OrderedListOutlined,
+  UnorderedListOutlined,
+  CopyOutlined,
+  ScissorOutlined,
+  PrinterOutlined,
+  LinkOutlined,
+  ShareAltOutlined,
+  LikeOutlined,
+  DislikeOutlined,
+  StarOutlined,
+  HeartOutlined,
+  FireOutlined,
+  ThunderboltOutlined,
+  BulbOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const actionIcons: IconMapping = {
+  plus: PlusOutlined,
+  edit: EditOutlined,
+  delete: DeleteOutlined,
+  save: SaveOutlined,
+  close: CloseOutlined,
+  check: CheckOutlined,
+  send: SendOutlined,
+  download: DownloadOutlined,
+  upload: UploadOutlined,
+  eye: EyeOutlined,
+  eyeInvisible: EyeInvisibleOutlined,
+  lock: LockOutlined,
+  unlock: UnlockOutlined,
+  reload: ReloadOutlined,
+  sync: SyncOutlined,
+  redo: RedoOutlined,
+  undo: UndoOutlined,
+  rollback: RollbackOutlined,
+  forward: ForwardOutlined,
+  backward: BackwardOutlined,
+  play: PlayCircleOutlined,
+  pause: PauseCircleOutlined,
+  stop: StopOutlined,
+  stepForward: StepForwardOutlined,
+  stepBackward: StepBackwardOutlined,
+  fastForward: FastForwardOutlined,
+  fastBackward: FastBackwardOutlined,
+  shrink: ShrinkOutlined,
+  arrowsAlt: ArrowsAltOutlined,
+  fullscreen: FullscreenOutlined,
+  fullscreenExit: FullscreenExitOutlined,
+  zoomIn: ZoomInOutlined,
+  zoomOut: ZoomOutOutlined,
+  compress: CompressOutlined,
+  expand: ExpandOutlined,
+  swap: SwapOutlined,
+  swapLeft: SwapLeftOutlined,
+  swapRight: SwapRightOutlined,
+  sortAscending: SortAscendingOutlined,
+  sortDescending: SortDescendingOutlined,
+  filter: FilterOutlined,
+  funnelPlot: FunnelPlotOutlined,
+  orderedList: OrderedListOutlined,
+  unorderedList: UnorderedListOutlined,
+  copy: CopyOutlined,
+  scissor: ScissorOutlined,
+  printer: PrinterOutlined,
+  link: LinkOutlined,
+  share: ShareAltOutlined,
+  like: LikeOutlined,
+  dislike: DislikeOutlined,
+  star: StarOutlined,
+  heart: HeartOutlined,
+  fire: FireOutlined,
+  thunderbolt: ThunderboltOutlined,
+  bulb: BulbOutlined,
+};
