+++ frontend-react/src/components/icons/communication.ts
@@ -1,0 +1,30 @@
+/**
+ * Communication Icons
+ * 
+ * Icons used for communication and user interaction.
+ */
+
+import {
+  MessageOutlined,
+  MailOutlined,
+  PhoneOutlined,
+  UserOutlined,
+  UserAddOutlined,
+  TeamOutlined,
+  LogoutOutlined,
+  LoginOutlined,
+  KeyOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const communicationIcons: IconMapping = {
+  message: MessageOutlined,
+  mail: MailOutlined,
+  phone: PhoneOutlined,
+  user: UserOutlined,
+  userAdd: UserAddOutlined,
+  team: TeamOutlined,
+  logout: LogoutOutlined,
+  login: LoginOutlined,
+  key: KeyOutlined,
+};
