+++ frontend-react/src/components/icons/data.ts
@@ -1,0 +1,42 @@
+/**
+ * Data Icons
+ * 
+ * Icons used for data visualization and tables.
+ */
+
+import {
+  BarChartOutlined,
+  TableOutlined,
+  BorderOutlined,
+  BorderInnerOutlined,
+  BorderOuterOutlined,
+  BorderTopOutlined,
+  BorderBottomOutlined,
+  BorderLeftOutlined,
+  BorderRightOutlined,
+  BorderVerticleOutlined,
+  BorderHorizontalOutlined,
+  RadiusUpleftOutlined,
+  RadiusUprightOutlined,
+  RadiusBottomleftOutlined,
+  RadiusBottomrightOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const dataIcons: IconMapping = {
+  barchart: BarChartOutlined,
+  table: TableOutlined,
+  border: BorderOutlined,
+  borderInner: BorderInnerOutlined,
+  borderOuter: BorderOuterOutlined,
+  borderTop: BorderTopOutlined,
+  borderBottom: BorderBottomOutlined,
+  borderLeft: BorderLeftOutlined,
+  borderRight: BorderRightOutlined,
+  borderVerticle: BorderVerticleOutlined,
+  borderHorizontal: BorderHorizontalOutlined,
+  radiusUpleft: RadiusUpleftOutlined,
+  radiusUpright: RadiusUprightOutlined,
+  radiusBottomleft: RadiusBottomleftOutlined,
+  radiusBottomright: RadiusBottomrightOutlined,
+};
