+++ frontend-react/src/components/icons/feedback.ts
@@ -1,0 +1,28 @@
+/**
+ * Feedback Icons
+ * 
+ * Icons used for user feedback and system status.
+ */
+
+import {
+  ExclamationCircleOutlined,
+  InfoCircleOutlined,
+  WarningOutlined,
+  CheckCircleOutlined,
+  LoadingOutlined,
+  SearchOutlined,
+  BellOutlined,
+  RobotOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const feedbackIcons: IconMapping = {
+  exclamation: ExclamationCircleOutlined,
+  info: InfoCircleOutlined,
+  warning: WarningOutlined,
+  checkCircle: CheckCircleOutlined,
+  loading: LoadingOutlined,
+  search: SearchOutlined,
+  bell: BellOutlined,
+  robot: RobotOutlined,
+};
