+++ frontend-react/src/components/icons/media.ts
@@ -1,0 +1,26 @@
+/**
+ * Media Icons
+ * 
+ * Icons used for media and file operations.
+ */
+
+import {
+  CameraOutlined,
+  VideoCameraOutlined,
+  AudioOutlined,
+  PictureOutlined,
+  FileOutlined,
+  FolderOutlined,
+  CloudOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const mediaIcons: IconMapping = {
+  camera: CameraOutlined,
+  video: VideoCameraOutlined,
+  audio: AudioOutlined,
+  picture: PictureOutlined,
+  file: FileOutlined,
+  folder: FolderOutlined,
+  cloud: CloudOutlined,
+};
