+++ frontend-react/src/components/icons/navigation.ts
@@ -1,0 +1,24 @@
+/**
+ * Navigation Icons
+ * 
+ * Icons used for navigation and layout purposes.
+ */
+
+import {
+  DashboardOutlined,
+  HomeOutlined,
+  MenuOutlined,
+  BarsOutlined,
+  MoreOutlined,
+  EllipsisOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const navigationIcons: IconMapping = {
+  dashboard: DashboardOutlined,
+  home: HomeOutlined,
+  menu: MenuOutlined,
+  bars: BarsOutlined,
+  more: MoreOutlined,
+  ellipsis: EllipsisOutlined,
+};
