+++ frontend-react/src/components/icons/system.ts
@@ -1,0 +1,38 @@
+/**
+ * System Icons
+ * 
+ * Icons used for system and technical operations.
+ */
+
+import {
+  SettingOutlined,
+  ToolOutlined,
+  AppstoreOutlined,
+  ApiOutlined,
+  DatabaseOutlined,
+  WifiOutlined,
+  SignalFilled,
+  PoweroffOutlined,
+  GlobalOutlined,
+  TranslationOutlined,
+  EnvironmentOutlined,
+  CalendarOutlined,
+  ClockCircleOutlined,
+} from '@ant-design/icons';
+import type { IconMapping } from './types';
+
+export const systemIcons: IconMapping = {
+  setting: SettingOutlined,
+  tool: ToolOutlined,
+  appstore: AppstoreOutlined,
+  api: ApiOutlined,
+  database: DatabaseOutlined,
+  wifi: WifiOutlined,
+  signal: SignalFilled,
+  poweroff: PoweroffOutlined,
+  global: GlobalOutlined,
+  translation: TranslationOutlined,
+  environment: EnvironmentOutlined,
+  calendar: CalendarOutlined,
+  clock: ClockCircleOutlined,
+};
